requirements.txt
cmake==3.25.0
dlib==19.19.0
face_recognition==1.3.0
pygame==2.2.0
imutils==0.5.4
Pillow==9.4.0
opencv-python==4.7.0.68
Flask==2.2.3
cv2==4.7.0.68
sklearn==1.0.2
numpy==1.21.6